cr: move opaque into helper
